Write images directly per pixel, omit on-the-fly colorbar rendering
The performance on the server was atrocious, and after some quick perf
poking it was clear that it was actually the rendering in pyplot that was
the bottleneck, not the data fetching that I was worried about.

The `pcolor` call was easy to replace, since on the client side we blow
up the pixels to our liking anyways. The colorbar however had to be redone
more ad-hoc, only showing the lower and upper limits on the client-side.
However, perf improved by more than 10x, with requests usually >10s taking
now ~1s typical.